requirements.txt
gpytorch==1.14
matplotlib==3.9.0
numpy==2.2.4
pandas==2.2.3
plotly==6.0.1
polars==1.26.0
pyro-ppl==1.9.1
torch==2.8.0
scikit-learn==1.6.1
scipy==1.15.2
tqdm==4.67.1
wandb==0.19.9